Move all code to module
