+++ .gitignore
@@ -8,4 +8,3 @@ user.cfg
 .idea/
 .vscode/
 venv/
-crypto_trading.db
