+++ binance_trade_bot/models/base.py
