+++ binance_trade_bot/models/coin.py
