+++ binance_trade_bot/models/coin_value.py
