+++ binance_trade_bot/models/pair.py
