+++ binance_trade_bot/models/scout_history.py
