+++ binance_trade_bot/models/trade.py
